Extract canonicalize into its own class
